==> src/2. alliance_construction.ipynb <==
# !git clone -b master https://github.com/StefanoNasini/Country-by-country-2022-2023-international-relationships
# !cd Country-by-country-2022-2023-international-relationships

import os
root = os.getcwd()
root = os.path.dirname(root)

# Bilateral trade agreement



with open(os.path.join(root, "raw_data", "bilateral_trade_data", "bilateral_trade_data.txt"), 'r', encoding='utf-8') as f:
    wholetext = f.read()

tradelist = wholetext.split("country")

import os

dir_agree = os.path.join(root, "trade_agreement")


with open(os.path.join(root, 'src', 'utils', 'tri2aka.json'), 'r') as f:
    tri2aka = eval(f.read())

with open(os.path.join(root, 'src', 'utils', 'orgs_d.json'), 'r') as f:
    orgs_d = eval(f.read())
#     f.read()

    
try:
    tri2aka.pop("ALN")
except KeyError:
    pass

tri2aka["EU"] = ["european union"]
tri2aka["LHT"] = ["lichtenstein"]

EU_members = orgs_d["EU"]

EU_memberset = set()
for (k, v) in EU_members.items():
    if v == 1:
        EU_memberset.add(k)

# remove \n start
for text in tradelist:
    text.strip()
tradelist = [text[1:] for text in tradelist if text[0] == "\n"]

groups = [text.split("\n") for text in tradelist]

agree = {}

os.chdir(os.path.join(root, "src", "utils"))
from aka import Nothing, Several, guess_country
os.chdir(root)

def record_rel(mother, child):
    if mother in agree.keys():
        if child not in agree[mother]:
            agree[mother].append(child)
    else:
        agree[mother] = [child]


show_process_country_text = False
def process_country_text(ctext):
    lines = ctext.split("\n")
    result = []
    mother_tri = None
    for line in lines:
        try:
            tri = guess_country(line, mother_tri)
            if mother_tri is None:
                mother_tri = tri
            result.append(tri)
            print("passed") if show_process_country_text else None
        except Nothing:
            print("not found ---------------") if show_process_country_text else None
            result.append(line)
        except Several:
            print("found several------------") if show_process_country_text else None
            result.append(line)
    return result

class Head(Exception):
    pass

def generate_second_processing(result):
    head_c = result[0]
    
    remain_aka_list = [head_c]
    added_text_list = [head_c]
    for result_item in result:
        if result_item == head_c:
            continue
        if result_item in tri2aka.keys():
            remain_aka_list.append(result_item)
        else:
            added_text_list.append(result_item)
    return remain_aka_list, added_text_list

ctext = tradelist[0]

tri2aka["EU"]

result = process_country_text(ctext)

tradelist

generate_second_processing(result)

agree = dict()
def update_done(done):
    mother = done[0]
    for child in done:
        if child == mother:
            continue
        record_rel(mother, child)
        record_rel(child, mother)


new_whole_text = []
for ctext in tradelist:
    result = process_country_text(ctext)
    done, to_be_processed = generate_second_processing(result)
    new_whole_text.append(to_be_processed)
    if len(done[0]) == 3:
        update_done(done)
    else:
        print(done)

for ctext in new_whole_text:
    print(ctext)

# record_rel("ARM", "ROU")
record_rel("CRI", "DOM")
show_guess_country = True
# guess_country("lichtenstein")

link_with_EU = """
Colombia
Costa Rica
Faroe Islands
Georgia
Israel
Japan
Jordan
Lebanon
Mexico
Peru
Serbia
Morocco
Singapore
South Korea
Switzerland
Tunisia
Turkey
United Kindom
Vietnam
Albania SAA (2009)
·       Algeria AA (2005)
·       Andorra CU (1991)
·       Bosnia and Herzegovina SAA (2015)
·   	Canada: Comprehensive Economic and Trade Agreement (provisionally applied)[23]
·       Central America AA
·       Chile AA (2003)
·       Egypt AA (2004)
·       Faroe Islands, autonomous entity of Denmark (1997)
·       Georgia AA/DCFTA (2014)
·   	Israel AA (2000)
·       Japan EU-Japan Economic Partnership Agreement (2019)
·       Jordan AA (2002)
·       Lebanon AA (2006)
·       Macedonia SAA (2004)
·   	Mexico AA (2000)
·       Montenegro SAA (2010)
·       Monaco CU (1958, Franco-Monegasque Treaties)
·       Morocco AA (2000)
·       Palestinian Authority interim AA (1997)
·       San Marino CU (2002)
·       Serbia SAA (signed 2008, entry into force on 1 September 2013)
·       South Africa AA (2000)
·       South Korea: European Union–South Korea Free Trade Agreement (entered in to force on 13 December 2015)[24]
·   	Switzerland FTA (1973)
·       Tunisia AA (1998)
·       Turkey CU (1996)
·       EU Overseas Countries and Territories[25]
·       Vietnam: European Union-Vietnam Free Trade Agreement (2020)
·       United Kingdom: European Union-United Kingdom Trade and Cooperation Agreement (entry into force 1 May 2021)
"""

process_country_text(link_with_EU)

# replace with trigraph manually
link_with_EU ={'ALB',
 'AND',
 'BIH',
 'CAN',
 'CHE',
 'CHL',
 'COL',
 'CRI',
 'DZA',
 'EGY',
 'FRO',
 'GBR',
 'GEO',
 'ISR',
 'JOR',
 'JPN',
 'KOR',
 'LBN',
 'MAR',
 'MCO',
 'MEX',
 'MNE',
 'PER',
 'SGP',
 'SMR',
 'SRB',
 'TUN',
 'TUR',
 'VNM',
 'ZAF'}

for euc in EU_memberset:
    for c in link_with_EU.union(EU_memberset):
        record_rel(euc, c)
        record_rel(c, euc)

# write down the result
if "constructed_network" not in os.listdir(root):
    os.mkdir("constructed_network")

os.chdir("constructed_network")
if "alliance_data" not in os.listdir():
    os.mkdir("alliance_data")
os.chdir(root)

with open(os.path.join(root, "constructed_network", "alliance_data", "trade_alliances.json"), 'w') as f:
    f.write(str(agree))

show_guess_country = True
# guess_country("dominican republic")

# Monetary alliances

trade_agreement = agree
agree = {}

# BRICS
ctexts = []
ctexts.append("Brazil, Russia, India, China, and South Africa".replace(',', '\n'))

# G7
ctexts.append("Canada, France, Germany, Italy, Japan, the United Kingdom ,and the United States".replace(',', '\n'))

# OPEC
ctexts.append("Iran, Iraq, Kuwait, Saudi Arabia, Venezuela, Qatar, Indonesia, Libya, United Arab Emirates, Algeria, Nigeria, Ecuador, Gabon, Angola, Equatorial Guinea, Congo.".replace(',', '\n'))

# ECOWAS
ctexts.append("Benin, Burkina Faso, Cabo Verde, Cote d'Ivoire, Gambia, Ghana, Guinea, Guinea-Bissau, Liberia, Mali, Niger, Nigeria, Senegal, Sierra Leone, Togo.".replace(',', '\n'))

# ASEAN
ctexts.append("Brunei, Cambodia, Indonesia, Laos, Malaysia, Myanmar, Philippines, Singapore, Thailand, Vietnam".replace(',', '\n'))

ctexts

ctextstri = []
for ctext in ctexts:
    l = []
    for c in ctext.split('\n'):
        try:
            tri = guess_country(c)
            l.append(tri)
        except Several:
            l.append(c)
        except Nothing:
            l.append(c)
    ctextstri.append(l)

ctextstri.append(list(EU_memberset))

ctextstri

# manually enter the member states
ctextstri = [['BRA', 'RUS', 'IND', 'CHN', 'ZAF'], ['CAN', 'FRA', 'DEU', 'ITA', 'JPN', 'GBR', 'USA'], ['IRN', 'IRQ', 'KWT', 'SAU', 'VEN', 'QAT', 'IDN', 'LBY', 'ARE', 'DZA', 'NGA', 'ECU', 'GAB', 'AGO', 'GNQ', 'COG'], ['BEN', 'BFA', 'CPV', 'CIV', 'GMB', 'GHA', 'GIN', 'GNB', 'LBR', 'MLI', 'NER', 'NGA', 'SEN', 'SLE', 'TGO'],['BRN', 'KHM', 'IDN', 'LAO', 'MYS', 'MYR', 'PHL', 'SGP', 'THA', 'VNM'],['BEL', 'IRL', 'LUX', 'POL', 'GRC', 'CYP', 'AUT', 'HUN', 'ROU', 'LTU', 'ESP', 'DNK', 'FIN', 'ITA', 'SVN', 'LVA', 'HRV', 'FRA', 'NLD', 'SWE', 'BGR', 'SVK', 'PRT', 'CZE', 'EST', 'DEU', 'MLT']]

for tris in ctextstri:
    for tri1 in tris:
        for tri2 in tris:
            if tri1 == tri2:
                continue
            record_rel(tri1, tri2)
            record_rel(tri2, tri1)

with open(os.path.join(root, "constructed_network", "alliance_data", "monetary_alliances.json"), 'w') as f:
    f.write(str(agree))

# Military alliance

agree = {}
ctexts = []

os.chdir(os.path.join(root, "src", "utils"))
from aka import name2tri
os.chdir(root)

# NATO 
NATO = "Belgium, Canada, Denmark, France, Iceland, Italy, Luxembourg, the Netherlands, Norway, Portugal, the United Kingdom, United States, Greec, Turkey; Germany; Spain; Czech Republic, Hungary, Poland; Bulgaria, Estonia, Latvia, Lithuania, Romania, Slovakia, Slovenia; Albania, Croatia, Montenegro; North Macedonia; Finland.".replace(',', '\n').replace(';', '\n')
ctexts.append(NATO)

# MNNA
mnna = "Argentina, Australia, Bahrain, Brazil, Colombia, Egypt, Israel, Japan, Jordan, Kuwait, Morocco, New Zealand, Pakistan, the Philippines, Qatar, South Korea, Thailand, and Tunisia".split(',')
for cname in mnna :
    try:
        c = name2tri[cname.lower().strip()]
    except KeyError:
        c = guess_country(cname, None, True)
    record_rel(c, "USA")
    record_rel("USA", c)

#Five eyes
five_eyes = "United States, United Kingdom, Canada, Australia, New Zealand".replace(',', '\n')

ctexts.append(five_eyes)

#CSTO
csto = " Armenia, Belarus, Kazakhstan, Kyrgyzstan, Russia, Tajikistan".replace(',', '\n')
ctexts.append(csto)

#SCO shanghai cooperation organization
sco = "China, Kazakhstan, Kyrgyzstan, Russia, Tajikistan, Uzbekistan, India, Iran, Pakistan, Belrus, Mongolia"
ctexts.append(sco.replace(',', '\n'))

#AUKUS 
aukus = "United States, Australia, United Kingdom".replace(',', '\n')
ctexts.append(aukus)

# Nordic defense cooperation
nordefco = "Denmark, Finland, Iceland, Norway, Sweden".replace(',', '\n')
ctexts.append(nordefco)

guess_country(" Solomon Islands.")
record_rel("CHN", "SLB")
record_rel("SLB", "CHN")

guess_country(" Sweden")
record_rel("SWE", "GBR")
record_rel("GBR", "SWE")

guess_country(" finland")
record_rel("FIN", "GBR")
record_rel("GBR", "FIN")

guess_country(" oman")
record_rel("OMN", "GBR")
record_rel("GBR", "OMN")

# Entente
ctexts.append("Egypt, Iraq, Jordan".replace(',', '\n'))

guess_country(" greece")
record_rel("JPN", "AUS")
record_rel("JPN", "GBR")
record_rel("GBR", "GRC")

record_rel("AUS", "JPN")
record_rel("GBR", "JPN")
record_rel("GRC", "GBR")

ctextstri

ctextstri = []
for ctext in ctexts:
    l = []
    for c in ctext.split('\n'):
        try:
            tri = guess_country(c) if c.lower().strip() not in name2tri else name2tri[c.lower().strip()]
            l.append(tri)
        except Several:
            l.append(c)
        except Nothing:
            l.append(c)
    ctextstri.append(l)


name2tri = {" India":"IND", " China":"CHN", " Equatorial Guinea":"GNQ", " Greec":"GRC", " Romania":"ROU", " the Netherlands":"NLD", " Nigeria":"NGA", " Niger":"NER", " Guinea-Bissau":"GNB", " Guinea":"GIN"}

for members in ctextstri:
    for (i, mem) in enumerate(members):
        if mem in name2tri.keys():
            members[i] = name2tri[mem]

ctextstri

for tris in ctextstri:
    for tri1 in tris:
        for tri2 in tris:
            if tri1 == tri2:
                continue
            record_rel(tri1, tri2)

with open(os.path.join(root, "constructed_network", 'alliance_data', 'military_alliances.json'), 'w') as f:
    f.write(str(agree))

# guess_country("guinea")

with open(os.path.join(root, "constructed_network", "alliance_data", "trade_alliances.json"), 'r') as f:
    trade = eval(f.read())
    
with open(os.path.join(root, "constructed_network", "alliance_data", "monetary_alliances.json"), 'r') as f:
    monetary = eval(f.read())
    
with open(os.path.join(root, "constructed_network", "alliance_data", "military_alliances.json"), 'r') as f:
    military = eval(f.read())

agree = {}
for d in [trade, monetary, military]:
    for tri, allies in d.items():
        for c in allies:
            record_rel(tri, c)

with open(os.path.join(root, "constructed_network", "alliances_network.json"), 'w') as f:
    f.write(str(agree))

os.chdir(os.path.join(root, "src", "utils"))
from aka import UN_tri
os.chdir(root)

# Inspect the countries that don't belong to UN
for k in agree.keys():
    if k in UN_tri:
        pass
    else:
        print(k)

# Remove countries outside UN for alliances
to_be_popped = []
for k, v in agree.items():
    if k == "CHN":
        print("Processing china-------")
        print(v)
    if k not in UN_tri:
        to_be_popped.append(k)
    else:
        for ele in v:
            print(f"processing {ele}") if k == "CHN" else None
            if ele not in UN_tri:
                v.remove(ele)
                print(f"{ele} removed from {k}")
        print("China processed", v) if k=="CHN" else None
        
for k in to_be_popped:
    agree.pop(k)

# store final alliance data into a pandas dataframe
import pandas as pd

alliances = pd.DataFrame(0, columns= sorted(list(UN_tri)), index=sorted(list(UN_tri)))
for k, v in agree.items():
    print(k)
    alliances.loc[k, v] = 1
    


alliances.to_csv(os.path.join(root, "constructed_network", "alliances.csv"))

"MAC" in agree["CHN"]

alliances
