==> src/9. Network_singleton_analysis.ipynb <==
# !git clone -b master https://github.com/StefanoNasini/Country-by-country-2022-2023-international-relationships
# !cd Country-by-country-2022-2023-international-relationships

import json
import os

root = os.path.dirname(os.getcwd())
import pandas as pd
os.chdir(os.path.join(root, "src", "utils"))
from aka import tri2name, tri2aka
os.chdir(os.path.join(root, "src"))


covotes = pd.read_csv(os.path.join(root, "constructed_network", "covotes.csv"), index_col=0)
contravotes = pd.read_csv(os.path.join(root, "constructed_network", "contravotes.csv"), index_col=0)

covotes_over80 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "covotes_over80.csv"), index_col=0)
covotes_over90 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "covotes_over90.csv"), index_col=0)
covotes_under20 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "covotes_under20.csv"), index_col=0)
covotes_under10 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "covotes_under10.csv"), index_col=0)
contravotes_under20 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "contravotes_under20.csv"), index_col=0)
contravotes_under10 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "contravotes_under10.csv"), index_col=0)
contravotes_over80 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "contravotes_over80.csv"), index_col=0)
contravotes_over90 = pd.read_csv(os.path.join(root, "constructed_network", "cocontravotes", "contravotes_over90.csv"), index_col=0)


alliances = pd.read_csv(os.path.join(root, "constructed_network", "alliances.csv"), index_col=0)
if "MAC" in alliances.columns:
    alliances.drop(columns="MAC", inplace=True)

alliances_zero = 1-alliances

trade_over80 = pd.read_csv(os.path.join(root, "constructed_network", "export_over80.csv"), index_col=0)
trade_over90 = pd.read_csv(os.path.join(root, "constructed_network", "export_over90.csv"), index_col=0)

trade_under20 = pd.read_csv(os.path.join(root, "constructed_network", "export_under20.csv"), index_col=0)
trade_under10 = pd.read_csv(os.path.join(root, "constructed_network", "export_under10.csv"), index_col=0)

trans_trade_over80 = pd.read_csv(os.path.join(root, "constructed_network", "import_over80.csv"), index_col=0)
trans_trade_over90 = pd.read_csv(os.path.join(root, "constructed_network", "import_over90.csv"), index_col=0)

trans_trade_under20 = pd.read_csv(os.path.join(root, "constructed_network", "import_under20.csv"), index_col=0)
trans_trade_under10 = pd.read_csv(os.path.join(root, "constructed_network", "import_under10.csv"), index_col=0)



df_names = [f'{a}{b}_{c}' for a in ["friends", "enemies"] for b in ["80", "90"] for c in ["bi", "tri"]]

# for df_name in df_names:
#     exec(f'{df_name} = pd.read_csv(os.path.join(root, "final_network", "{df_name}.csv"), index_col=0)')

friends80_bi = pd.read_csv(os.path.join(root, "constructed_network", "final_network", "friends80_bi.csv"), index_col=0)
friends80_tri = pd.read_csv(os.path.join(root,"constructed_network",  "final_network", "friends80_tri.csv"), index_col=0)
friends90_bi = pd.read_csv(os.path.join(root, "constructed_network", "final_network", "friends90_bi.csv"), index_col=0)
friends90_tri = pd.read_csv(os.path.join(root,"constructed_network",  "final_network", "friends90_tri.csv"), index_col=0)
enemies80_bi = pd.read_csv(os.path.join(root, "constructed_network", "final_network", "enemies80_bi.csv"), index_col=0)
enemies80_tri = pd.read_csv(os.path.join(root,"constructed_network",  "final_network", "enemies80_tri.csv"), index_col=0)
enemies90_bi = pd.read_csv(os.path.join(root, "constructed_network", "final_network", "enemies90_bi.csv"), index_col=0)
enemies90_tri = pd.read_csv(os.path.join(root,"constructed_network",  "final_network", "enemies90_tri.csv"), index_col=0)

import os
os.chdir(root)



df_names = [f'{a}{b}_{c}' for a in ["friends", "enemies"] for b in ["80", "90"] for c in ["bi", "tri"]]

for df_name in df_names:
    print(f'{df_name} = pd.read_csv(os.path.join(root, "final_network", "{df_name}.csv"), index_col=0)')


import numpy as np
import networkx as nx

def analyze_graph(adj_matrix):
    # Create a graph from the adjacency matrix
    G = nx.Graph(adj_matrix)
    
    # Number of singletons (nodes with degree 0)
    num_singletons = sum(1 for node in G.nodes() if G.degree(node) == 0)
    
    # Number of edges
    num_edges = G.number_of_edges()
    
    # Node with the largest degree
    largest_degree_node = max(G.nodes(), key=lambda node: G.degree(node))
    largest_degree = G.degree(largest_degree_node)
    
    # Node with the least degree
    least_degree_node = min(G.nodes(), key=lambda node: G.degree(node))
    least_degree = G.degree(least_degree_node)
    
    return num_singletons, num_edges, largest_degree_node, largest_degree, least_degree_node, least_degree

def complete_analyze_graph(dataframe, print_=False):
    (ns, ne, Ldn, Ld, ldn, ld) = analyze_graph(dataframe.to_numpy())
    print(f"""number of singletons={ns}\n 
            number of edges={ne}\n
            Node w/ largest degree ={dataframe.columns[Ldn]}\n
            country name={tri2name[dataframe.columns[Ldn]]}\n
            Largest degree = {Ld}\n
            Node w/ least degree = {dataframe.columns[ldn]}\n
            country name={tri2name[dataframe.columns[ldn]]}\n
            least degree = {ld}""") if print_ else None
    return ns, ne, Ld, dataframe.columns[Ldn], tri2name[dataframe.columns[Ldn]], ld, dataframe.columns[ldn], tri2name[dataframe.columns[ldn]]




table=pd.DataFrame(columns=["n_singletons", "n_edges", 
                            "largest_degree", "largest_degree_tri", "largest_degree_name", 
                            "least_degree", "least_degree_tri", "least_degree_name"])

for df_name in df_names:
    print(df_name, end='---------------------\n') if False else None
    df = eval(df_name)
    table.loc[df_name, :] = complete_analyze_graph(df)

for df_name in ['covotes_over80', 'covotes_over90', 'covotes_under20', 'covotes_under10', 
                'contravotes_under20', 'contravotes_under10', 'contravotes_over80', 'contravotes_over90']:
    table.loc[df_name.strip(), :] = complete_analyze_graph(eval(df_name))


for df_name in """trade_over90 trade_over80 trade_under20 trade_under10""".split(' '):
    table.loc[df_name, :] = complete_analyze_graph(eval(df_name))

for df_name in """trade_over90 trade_over80 trade_under20 trade_under10""".split(' '):
    df_name = "trans_"+df_name
    table.loc[df_name, :] = complete_analyze_graph(eval(df_name))

table.loc["alliances", :] = complete_analyze_graph(alliances)
table.loc["alliances_zero", :] = complete_analyze_graph(alliances_zero)


table["least_degree_"] = str(table["least_degree"])

# formatting least and largest degree
for index in table.index:
    table.loc[index, "least_degree_"] = f'{str(table.loc[index, "least_degree"])} ({str(table.loc[index, "least_degree_tri"])})'
    table.loc[index, "largest_degree_"] = f'{str(table.loc[index, "largest_degree"])} ({str(table.loc[index, "largest_degree_tri"])})'

table

def overleaf(networks, columns_to_show=None, networks_to_show=None):
    columns = ["n_singletons", "n_edges", "largest_degree_", "least_degree_"]
    if columns_to_show is None:
        columns_to_show = ["singletons", "edges", "degr. max", "degr. min"]
    if networks_to_show is None:
        networks_to_show = networks
    df = table.loc[networks, columns].transpose()
    n, m = df.shape
    print('& '+' & '.join(networks_to_show)) 
    print("\\clines{2-9}")
    for i in range(n):
        print(columns_to_show[i], end=' & ')
        for j in range(m):
            print(df.iloc[i, j], end='')
            print(" & ", end='') if j < m-1 else None
        print(r'\\')
        
    print()
    print(table.loc[networks, ["largest_degree_tri", "largest_degree_name", "least_degree_tri", "least_degree_name"]])
        
    
        


# cocontravotes over 80
networks = ["covotes_over80", "contravotes_over80", "trade_over80", "trans_trade_over80","covotes_over90", "contravotes_over90", "trade_over90","trans_trade_over90" ,"alliances"]
overleaf(networks)

# cocontravotes under 20 and under 10

networks = ["covotes_under20", "contravotes_under20", "trade_under20", "trans_trade_under20","covotes_under10", "contravotes_under10", "trade_under10", "trans_trade_under10", "alliances_zero"]
networks_to_show = ["covotes", "contravotes", "exports", "imports","covotes", "contravotes", "exports", "imports", "allinaces"]

overleaf(networks, networks_to_show = networks_to_show)

# friend-enemy 80 and 90
networks = ["friends80_tri", "enemies80_tri", "friends90_tri", "enemies90_tri"]
networks_to_show = ["friends", "enemies"]*2
overleaf(networks, networks_to_show=networks_to_show)

# cocontravotes under 20 and under 10

networks = ["covotes_under20", "contravotes_under20", "trade_under20","covotes_under10", "contravotes_under10", "trade_under10", "alliances_zero"]
networks_to_show = ["covotes", "contravotes", "trade","covotes", "contravotes", "trade", "allinaces"]

overleaf(networks, networks_to_show = networks_to_show)

# cocontravotes over 90
networks = ["covotes_over90", "contravotes_over90", "trade_over90", "alliances"]
overleaf(networks)

# friend-enemy 80 and 90
networks = ["friends80_bi", "enemies80_bi", "friends90_bi", "enemies90_bi", "friends80_tri", "enemies80_tri", "friends90_tri", "enemies90_tri"]
networks_to_show = ["friends", "enemies"]*4
overleaf(networks, networks_to_show=networks_to_show)

# friend-enemy 90
networks = ["friends90_bi", "enemies90_bi", "friends90_tri", "enemies90_tri"]
overleaf(networks)
